% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_default_no_proxy}
\alias{get_default_no_proxy}
\title{Get Default No Proxy String}
\usage{
get_default_no_proxy()
}
\value{
The default no_proxy string, or NULL if not set.
}
\description{
Retrieves the default no_proxy string. Checks the R option \code{proxyr.no_proxy} first,
then the \code{NO_PROXY} environment variable.
}
\examples{
get_default_no_proxy()
}
