% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{list_proxy_credentials}
\alias{list_proxy_credentials}
\title{List Proxy Credentials}
\usage{
list_proxy_credentials()
}
\value{
A data frame of keys.
}
\description{
Lists all proxy credentials stored by this package (or matching services).
Since keyring stores by service name, this lists all services, or we relies on user knowing.
}
\details{
Note: This function lists ALL keys in the keyring if we don't filter.
To make this safer, we might only list those that look like urls or just wrap key_list.
}
\examples{
\dontrun{
list_proxy_credentials()
}
}
