% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{remove_proxy_credentials}
\alias{remove_proxy_credentials}
\title{Remove Proxy Credentials}
\usage{
remove_proxy_credentials(proxy_url)
}
\arguments{
\item{proxy_url}{The URL of the proxy server.}
}
\description{
Removes the stored credentials for a specific proxy server.
}
