% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{set_default_no_proxy}
\alias{set_default_no_proxy}
\title{Set Default No Proxy String}
\usage{
set_default_no_proxy(no_proxy)
}
\arguments{
\item{no_proxy}{The no_proxy string (e.g., "localhost,127.0.0.1").}
}
\description{
Sets the default no_proxy string to be used by \code{\link{with_proxy}}.
}
