% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{set_default_proxy_url}
\alias{set_default_proxy_url}
\title{Set Default Proxy URL}
\usage{
set_default_proxy_url(proxy_url)
}
\arguments{
\item{proxy_url}{The URL of the proxy server (e.g., "http://proxy.example.com:8080").}
}
\description{
Sets the default proxy URL to be used by \code{\link{with_proxy}}.
}
