% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{set_proxy_credentials}
\alias{set_proxy_credentials}
\title{Set Proxy Credentials}
\usage{
set_proxy_credentials(proxy_url, username, password)
}
\arguments{
\item{proxy_url}{The URL of the proxy server (e.g., "http://proxy.example.com:8080").
This is used as the service name in keyring.}

\item{username}{The username for the proxy.}

\item{password}{The password for the proxy.}
}
\description{
Stores the username and password for a specific proxy server in the Windows
Credential Manager using the keyring package.
}
\section{Security Warning}{

You should \strong{never} hardcode your username and password in your R scripts
or commit them to version control. This function is intended to be run
\strong{once} interactively to store your credentials securely in the Windows
Credential Manager. Once stored, you do not need to run this function again
unless your credentials change.
}

\examples{
\dontrun{
# Run this ONCE interactively, do not include in your scripts:
set_proxy_credentials("http://my.proxy.com:8080", "myuser", "mypass")
}
}
